test/slush.js: Refactor runSlush, capture stderr
Eliminates a lot of common boilerplate from each test case, and fixes
other issues. Also adds assertions to ensure events are triggered as
expected.

Previously any assertion errors would technically go uncaught. The more
proper way to report them in async tests is to set up a try/catch and
pass the error to `done()`.

Both stdout and stderr are now captured, and the "should fail when
running a generator without slushfile" now fails. Previously, when slush
failed in this test, stdout was empty while stderr was ignored due to a
failure in liftoff (`cli.launch()`). The `data` variable used to capture
stdout was empty, which caused `should.match` to always pass.

All output is now written to stderr when a test fails to facilitate
stack trace debugging (as with the `cli.launch()` failure).