package.json,bin/slush.js: Bump to liftoff 2.5.0
Fixes the "should fail when running a generator without slushfile" test
case.

As noted in the commit that updated runSlush() from test/slush.js and
caused the failure, slush was actually failing without the expected "No
slushfile found" message. It was failing with a stack trace that
started with:

  path.js:39
      throw new ERR_INVALID_ARG_TYPE('path', 'string', path);
    ^

  TypeError [ERR_INVALID_ARG_TYPE]: The "path" argument must be of type
  string. Received type object
      at assertPath (path.js:39:11)
      at Object.extname (path.js:1375:5)
      at Liftoff.buildEnvironment (/Users/mbland/src/mbland/slush/node_modules/liftoff/index.js:111:50)
      at Liftoff.launch (/Users/mbland/src/mbland/slush/node_modules/liftoff/index.js:154:22)
      at Object.<anonymous> (/Users/mbland/src/mbland/slush/bin/slush.js:57:5)

This points to a bug in the previous version of liftoff whereby the the
search for `slushfile.js` found nothing, resulting in a `configPath` of
`null`. liftoff passed this `null` to `Object.extname`, resulting in
this stack trace.

liftoff v2.5.0 doesn't have this problem, but it does have a slightly
different API. With the changes to `bin/slush.js`, this test case now
passes along with the rest of the suite.